Refactor the tests structure (#1157)
A lot of functions did belong to the main test package, often added to
the bottom of the file. Many functions were used both by l2 and bgp
tests, but living in one file or the other.

Now that the tests are settled a bit, we add a set of new packages, with
the intent to provide a better home for the code shared by both L2 and
bgp tests.

We also split the bgp and l2 tests in two separate packages, to enforce
the use of packages when the code is shared between the two.
